refactor(pds): move helpers into util
- Move PDS helper functions out of driver.go into util.go.

- Remove the PDS driver test file from the tracked tree while keeping local tests available.
